#37 consistent usage of 'yaml' over 'yml' in code, with the exceprion of file extension, which remains as '.yml'
