+++ src/test/scala/com/atomist/parse/java/spring/ApplicationYamlAssertions.scala
@@ -1,0 +1,11 @@
+package com.atomist.parse.java.spring
+
+import com.atomist.source.ArtifactSource
+
+object ApplicationYamlAssertions {
+
+  val ApplicationYamlFilePath = "src/main/resources/application.yml"
+
+  def hasApplicationYaml(as: ArtifactSource): Boolean =
+    as.findFile(ApplicationYamlFilePath).isDefined
+}
